Moved session related stuff in webserver to subpackage "sessions"
