+++ src/main/java/de/igslandstuhl/database/server/webserver/PostRequestHandler.java
@@ -36,6 +36,8 @@ import de.igslandstuhl.database.server.webserver.requests.HttpHandler;
 import de.igslandstuhl.database.server.webserver.requests.PostRequest;
 import de.igslandstuhl.database.server.webserver.responses.HttpResponse;
 import de.igslandstuhl.database.server.webserver.responses.PostResponse;
+import de.igslandstuhl.database.server.webserver.sessions.Session;
+import de.igslandstuhl.database.server.webserver.sessions.SessionManager;
 import de.igslandstuhl.database.utils.JSONUtils;
 import de.igslandstuhl.database.utils.ThrowingConsumer;
